msm: ipa: delete immediate command construct functions
The only caller of the construct functions in the immediate command
table was ipahal_construct_imm_cmd(), and that was removed by the
previous patch.  Remove the construct field in the immediate command
structure, and drop the function references that were previously
stored there.  The functions are unreferenced as a result and must
be removed.

Each of these functions has a parameter block structure associated
with it, and once the functions are gone those types are no longer
used, so they can go away as well.  Add a little documentation about
the payload allocatiion function parameters.

In addition, the ipahal_imm_cmd_table_dma structure is also removed
because it was never referenced in the IPA lite code

Change-Id: I67cfdaaf8d1e55f5c9e8f44cb0231522853244e8
Signed-off-by: Alex Elder <elder@codeaurora.org>