msm: ipa: kill ipa3_tag_destroy_imm()
There is already a callback function whose purpose is to free the
payload for an immediate command, ipa3_destroy_imm().  It's private
though, and perhaps for that reason there is an identical function
named ipa3_tag_destroy_imm().  Get rid of that function and use
ipa3_destroy_imm() instead (after making it public).

Change-Id: Ic253bd926abb113f8e5b38966eafbfe5127906a2
Signed-off-by: Alex Elder <elder@codeaurora.org>