msm: ipa: return earlier in ipa3_assign_policy() for APPS_LAN_CONS
In ipa3_assign_policy(), if for consumer clients, there is a large
if/else block, followed by a return.  Add a return statement to the
end of the "if" block (for IPA_CLIENT_APPS_LAN_CONS), allowing the
"else" block to be shifted out (that shifting is the only change to
this patch makes to the contents of that block).

Change-Id: Id6b38299c299488acfe6f3d99bd5a4b7a37da10c
Signed-off-by: Alex Elder <elder@codeaurora.org>