msm: ipa: ep->sys is non-null for APPS_CMD_PROD
Two spots check whether the system context associated with the
endpoint for the APPS_CMD_PROD client is null.  It will not be.
An endpoint's system context is set in ipa3_setup_sys_pipe(), which
for the APPS_CMD_PROD pipe is called by ipa3_setup_apps_pipes().  If
the allocation of the system context fails, ipa3_setup_sys_pipe()
will fail, and for the APPS_CMD_PROD pipe ipa3_post_init() will
also fail and the IPA driver will not be functional.  So we'd never
reach these functions checking for null if it were...

Change-Id: Icf87ab55c5a17a4508ec656e355a86d8a71985bc
Signed-off-by: Alex Elder <elder@codeaurora.org>