msm: ipa: look up endpoint context later when sending a command
We know the endpoint lookup for the APPS_CMD_PROD pipe will be good
when sending an immediate command in ipa3_send_cmd_timeout() or
ipa3_send_cmd().  We can do the lookup at any time, so do so just
before it's needed in both functions--just before the call to
ipa3_send().

Change-Id: Ie86dddb05f6d0c479834612b74fa2b8b51f30609
Signed-off-by: Alex Elder <elder@codeaurora.org>