msm: ipa_dp: fix ipa3_tx_dp() comments
Only the WAN producer pipe path calls ipa3_tx_dp().  There have
been other changes to that function as part of cleaning it up for
upstreaming, and much of the information in the comment atop the
function no longer applies.  Get rid of almost all of that.

The "dst" parameter really identifies the source pipe.  In the full
IPA code it can sometimes represent a source and sometimes a
destination.  Rename it "client" so it doesn't suggest either one.

Get rid of the confusing comments about how "dst" was supposed to be
interpreted for various scenarios.  We have only one now, and it
doesn't really need any special explanation.

Change-Id: Ife939e8bde1125b06cebc7fe82db1e6e3d8dda6f
Signed-off-by: Alex Elder <elder@codeaurora.org>