msm: ipa: client is good and valid in ipa3_tx_dp()
The only caller of ipa3_tx_dp() passes APPS_WAN_PROD as the client
id.  We can see by inspection that this client has an entry in the
ipa3_ep_configuration[][] array, and its entry is valid, and has an
in-range index value (it is 8), so "gsi_ep" will be non-null.

Change-Id: I5869f5475e738cca14b8431adee88dca8058ed25
Signed-off-by: Alex Elder <elder@codeaurora.org>