msm: ipa: no extra descriptors required in ipa3_tx_dp()
Earlier, if a metadata buffer was supplied as a parameter to
ipa3_tx_dp(), two additional descriptors might be required to
encode an skb into an IPA request.  That parameter wasn't used in
IPA lite and was removed, and consequently we no longer need
anything more than one descriptor for the linear portion and an
additional descriptor for any fragments in the skb.  Update the
code accordingly.  The (presumably common) case of no fragments
is handled using a single local descriptor variable (rather than
the array of three used previously).

Change-Id: I04d3b01e9335a437b8f985955e666362f0c0c73d
Signed-off-by: Alex Elder <elder@codeaurora.org>