msm: ipa: clnt_hdl is good in ipa3_rx_poll()
The only caller of ipa3_rx_poll() is ipa3_rmnet_poll(), and the only
way that function will be called is as the poll function supplied to
netif_napi_add() when initialized by ipa3_wwan_probe().  The value
of "clnt_hdl" is the ipa3_to_apps_hdl, which must have been by
handle3_ingress_format().  If it had not, the endpoint would never
cause apps_ipa_packet_receive_notify(), and that's what eventually
leads to ipa3_rmnet_poll() being called.  So we can assume the
"clnt_hdl" value supplied to ipa3_rx_poll() is valid.

Change-Id: I7531090e5b78d323a4c82761096912b777ba9833
Signed-off-by: Alex Elder <elder@codeaurora.org>