msm: ipa: kill ipa3_wq_rx_avail()
The only times ipa3_wq_repl_rx() is called is as a work function
associated with a received packet.  The work is initialized in four
places, all looking like this:
    INIT_WORK(&rx_pkt->work, ipa3_wq_rx_avail);
But nothing ever schedules that work to run, so the function is
unused.  Delete these calls to INIT_WORK(), and the function, and
the work structure embedded in the ipa3_rx_pkt_wrapper structure.

Change-Id: I7eb3335fec2c4b47632212a82c21924ee202b67b
Signed-off-by: Alex Elder <elder@codeaurora.org>