msm: ipa: don't use common event pipe for APPS_WAN_PROD
A common event pipe is allocated to allow producer pipes (which
produce relatively light traffic on their corresponding event pipe)
to share.  Currently the only client that is set up to use the common
pipe is APPS_WAN_PROD.  So it's not shared at all, and we get no
real benefit from special handling.  Switch the APPS_WAN_PROD so it
no longer uses the shared pipe for events and instead allocates a
new (exclusive) one.

Change-Id: If7d89cc46a09426dcbf9f3a9926602e7a5d13a8e
Signed-off-by: Alex Elder <elder@codeaurora.org>