msm: ipa: kill ipa3_sys_context->db_timer
The db_timer field of the ipa3_sys_context structure is now
initialized but never started.  It was used only for scheduling a
timeout when the common producer event channel was used, but the
previous patch got rid of that.  Delete the field, and delete
ipa3_ring_doorbell_timer_fn(), which was only used by the timer.

Change-Id: I14d5c76d0f3b60ff12e57856969d20af6c14fdf6
Signed-off-by: Alex Elder <elder@codeaurora.org>