msm: ipa: kill ipa3_send_nop_desc()
The only way ipa3_send_nop_desc() is called is as a work function,
set up for the APPS_WAN_PROD channel, in ipa3_assign_policy().  That
is, it would be called as a result of queue_work(sys->wq, &sys->work)
being called for the WAN producer endpoint.

There were (until recent commits) 4 places where a system context's
work structure was activated (by queue_work()):
  - ipa_gsi_irq_rx_notify_cb() (twice).  This is the xfer_cb
    function used for consumer (CONS) channels, and will not be
    called for the WAN producer endpoint.
  - ipa3_gsi_poll_after_suspend().  This is called in two places
    by ipa3_suspend_apps_pipes(), involving the APPS_LAN_CONS and
    APPS_WAN_CONS pipes.  So again, this function will never be called
    for the WAN producer channel.
  - ipa3_ring_doorbell_timer_fn().  This function would be called
    when a timer expired.  The timer was started in ipa3_send() only
    for pipes using the common event ring.  The WAN producer no
    longer uses the common event ring, so this function will not be
    called for for that channel.  (And there were no other users, so
    this function is now gone.)
  - ipa3_send_nop_desc().  This is (was) the only remaining place
    that would activate the work function for the WAN producer pipe.
    It would only do this if a memory allocation request failed.

Nothing will do an (initial) activation of the work structure in
the system context of the WAN producer pipe.  So ipa3_send_nop_desc()
will never get called, and it can go away.

Add a comment next to the "work" field definition in the system
context structure indicating it's only used for consumer pipes.

Change-Id: I8808f2b845b24682d66cfe12338af735b0cd02ea
Signed-off-by: Alex Elder <elder@codeaurora.org>