msm: ipa: rename ipa_assert() to be ipa_bug()
The way ipa_assert() is used is different from the way assertions
are normally defined.  Typically an assertion includes a condition,
and defines an assumption of the code that can optionally be verified
at runtime.  The way ipa_assert() is used in the IPA code is really
more like a BUG() call, in that it's called unconditionally to force
a crash because something is wrong.  Failed assertions might cause
BUG() calls, but their purpose is different.

Another important difference between assertions and BUG() calls is
that an assertion (including the condition it asserts) must not
contain code that has side-effects, because it can be compiled out.
A BUG() (or BUG_ON()) call may not always crash the system, but it
will always be compiled in, and can be assumed to halt the current
execution path.

Rename ipa_assert() to be ipa_bug(), to better indicate the purpose
it serves.

Change-Id: I7a243162a68157a848b82160921c315ff08d7a38
Signed-off-by: Alex Elder <elder@codeaurora.org>