msm: ipa: make more use of ipa_bug_on()
Replace a lot of calls to ipa_bug() within conditional blocks with
ipa_bug_on(condition).  The only places that use ipa_bug() are in
switch statements (where the condition isn't as easily available) or
in places where the condition alone doesn't supply sufficient
information to diagnose a problem (in which case some information
might be dumped before calling ipa_bug()).

When ipa_bug_on() is supplied a function returning an error code,
use the form:
    ipa_bug_on(func() != 0);
rather than:
    ipa_bug_on(func());

I believe this "reads better," as in, "it's a bug if func returns an
error" rather than "it's a bug when funct is called."  (This is very
subjective, but I prefer it.)

Change-Id: Ifdf02ebafcf42b666f45e35ccdc7e9dbe6bc8a93
Signed-off-by: Alex Elder <elder@codeaurora.org>