msm: ipa: introduce ipa_assert()
Introduce a (new) macro ipa_assert(), which asserts that the
condition it is supplied is true.  The purpose of this differs
from ipa_bug_on(), which is meant to check on conditions (such
as unexpected results supplied by hardware) that should be treated
as unrecoverable errors.  An assertion is meant to communicate
assumptions of the code to the reader; a failed assertion indicates
that there is a software bug.

An example would be asserting that a function parameter value is in
range--it tells the reader what the function expects, but in practice
this should not have to be verified at runtime.  (A developer using
the function is responsible for ensuring the arguments supplied are
valid.)

Change-Id: I2c52c3b49aac165e9bd9a9d9ffe3a0dfe58af0f0
Signed-off-by: Alex Elder <elder@codeaurora.org>