msm: ipa: create ipa/active_clients/log
Create an IPA debugfs file "log" in the "active_clients" directory.
This file dumps the contents of the active clients log.  This
implementation leverages the existing sequential file support,
allowing an iterator to be used to have a single log buffer to be
printed at a time.  The result avoids the need for allocating a
special buffer to format the active clients information into, and
avoids the need for ipa3_active_clients_log_print_buffer() (found
in "ipa.c" though only used for this purpose).

Change-Id: Ifd04b3df6161662d96c01bc7420dd490c808ced2
Signed-off-by: Alex Elder <elder@codeaurora.org>