msm: ipa: kill ipa3_active_clients_log_print_buffer()
The "ipa/active_clients/log" debugfs file now dumps the contents of
the active clients log, so the existing "ipa/active_clients_log"
file no longer needs to do that.  This removes the need for
ipa3_active_clients_log_print_buffer() to format the content of
the log into a provided buffer, so get rid of it.

Change-Id: Ib2c0f7801a2eb47f10c8a2c13388e869f5a4cf0b
Signed-off-by: Alex Elder <elder@codeaurora.org>