msm: ipa: stop pre-filling log records
Every active client log record is filled with underscores before
formatting the actual information to be logged.  There is no point
in doing this.  The sprintf() call will NUL-terminate the formatted
string, and the log entry, when printed, will not show anything
beyond that terminator.

Stop filling the entries with underscores.

Change-Id: I63d8906ad31f87553ff0d4b95843a701579190f9
Signed-off-by: Alex Elder <elder@codeaurora.org>