msm: ipa: kill IPA3_ACTIVE_CLIENT_LOG_TYPE_*
Four symbols defined in "ipa.c" are redundantly defined.  Their
names and values are aligned with the ipa_active_client_log_type
enumerated type values, but they're only used in the code that
formats an active client log entry into a buffer.  The code that
creates those entries use the enumerated type instead.  (This is an
error.)  Get rid of the redundant set of symbols and use the
enumerated type only.

Change-Id: Ib773b07a0d5934670eee59bfe67e2f8720ae6ff5
Signed-off-by: Alex Elder <elder@codeaurora.org>