msm: ipa: introduce active_client_get()
Create a new function active_client_get() which looks up an active
client entry by name, and if it's found, returns it.  If it not
found, a new one is allocated and initialized, and a pointer to that
is returned.  A new entry has a count value of 0.  NULL is returned
if allocating memory for a new entry fails.

Using this function in ipa3_active_clients_log_mod() simplifies it
quite a bit.  Allocate the memory using GFP_ATOMIC--because that's
the way it happens anyway.

Change-Id: Ic298a5c76f3026f25b37eaa45744f3edffb64426
Signed-off-by: Alex Elder <elder@codeaurora.org>