msm: ipa: kill ipa3_context->tag_process_before_gating
Nothing ever sets tag_process_before_gating to true, so its value
will always be false.  Delete the field.  There is one spot where
its value is read, and if true, it's set to false.  Delete those two
lines of code and replace the conditional test with false--because
it would never had occurred.

Change-Id: Ieccc260e355b952129fb8ee777e20c3c27c3471a
Signed-off-by: Alex Elder <elder@codeaurora.org>