msm: ipa: kill ipa3_start_tag_process()
The only place ipa3_start_tag_process() is called is as a work
function for ipa3_tag_work.  But that is never queued to be run
because the block in which it's queued is conditional on false.

Delete the call to queue_work() that will never happen, and delete
ipa3_start_tag_process() and the work structure that used it.

Change-Id: I9f06b5899ccdd9c645bb3986802babbcd0d0cd4f
Signed-off-by: Alex Elder <elder@codeaurora.org>