msm: ipa: no need to suspend pipes in ipa3_pre_init()
In the error path of ipa3_pre_init(), ipa3_suspend_apps_pipes() is
called before calling ipa3_disable_clks().  There is no need to
suspend pipes in this case.

When ipa3_suspend_apps_pipes() is called the LAN and WAN consumer
pipes are suspended.  If a pipe is not marked valid at the time
it is suspended, it is not considered an error--the operation is
simply skipped.

A pipe is marked valid by ipa3_setup_sys_pipe(), and until this has
been done successfully the pipe is not valid.

The LAN consumer pipe is set up near the end of ipa3_setup_apps_pipes().
That function isn't called until ipa3_post_init(), which occurs well
after ipa3_pre_init().  So the call to ipa3_suspend_apps_pipes() can
be removed from the error path of ipa3_pre_init().

Change-Id: Ic819c3beedbaccb70c3e3762c2ac1276259e8f29
Signed-off-by: Alex Elder <elder@codeaurora.org>