msm: ipa: fix bug in ipa3_isr()
The ipa3_dec_client_disable_clks() call at the end of ipa3_isr()
should use the non-blocking version of that function.

The reason is that there's a chance that between the call to
ipa3_inc_client_enable_clks_no_block() earlier in the function and
the ipa3_dec_client_disable_clks() at the end, the only other
reference got dropped.  In that case, the reference dropped at the
end of ipa3_isr() will be the last one, which leads to disabling
activity that blocks.  That can't happen in an interrupt routine.

Change-Id: I136cd2b2e1a924621c632b8a45d29a090b981670
Signed-off-by: Alex Elder <elder@codeaurora.org>