msm: ipa: clarify something in ipa3_rx_poll()
Move the call to IPA_ACTIVE_CLIENTS_PREP_SPECIAL() in ipa3_rx_poll()
to be just before the use of the log structure it initializes.  Move
the definition of that log structure to be within the same block.

Add a comment to indicate that the matching client increment and
clock enable call is in ipa_gsi_irq_rx_notify_cb(), and add a
similar comment in that function referring back to ipa3_rx_poll().

Change-Id: Ie3915b02772fb4f62e3d91c8ec4716a5f1273648
Signed-off-by: Alex Elder <elder@codeaurora.org>