msm: ipa: drop int_ctx argument in ipa3_active_clients_log_mod()
The "int_ctx" argument to ipa3_active_clients_log_mod() is supposed
to indicate whether the caller is running in interrupt context.  But
it's never used--the function passes GFP_ATOMIC when allocating
memory regardless of what is passed.  Drop the argument; we can
figure out how and when to use GFP_KERNEL at some future date.

Change-Id: I5e361a80344a195932127ce029f2b07e566a032a
Signed-off-by: Alex Elder <elder@codeaurora.org>