msm: ipa: always disable clocks asynchronously
A non-blocking clock disable call arranges to disable the clock in
work queue context if the last reference is dropped.  There's very
little benefit to disabling clocks immediately for a blocking clock
disable call; doing the disable requires a mutex, and it's possible
the dropped reference is no longer the last once the mutex is
acquired.

Just have the "blocking" clock disable call use the no-blocking
version, and have the disable always happen in a workqueue.

Change-Id: I5b91cc29eb15618532ff25e8e78130f91218bee0
Signed-off-by: Alex Elder <elder@codeaurora.org>