msm: ipa: introduce ipa_client_add_first()
This is the first of several patches that rework (and rename) client
reference counting and clock management for IPA.

We will have four routines for adding or removing clients.  Client
reference counts determine whether IPA hardware is active (and when
it should be activated or brought down).

  - void ipa_client_add();
	Add a new client reference, blocking if necessary while
	clocks are enabled and pipes are brought up.

  - bool ipa_client_add_additional();
	Add a new client, but only if an existing client has IPA
	hardware is active (clocks enabled and pipes up).  Returns
	true if a reference was taken.  Returns false (without
	activating hardware and without taking are reference) if the
	IPA hardware was not active.

  - void ipa_client_remove();
	Removes an existing client reference, arranging for the
	pipes to be brought down and clocks stopped when the last
	reference is removed.  Does not block.

  - void ipa_client_remove_wait();
	Removes an existing client reference, and if this was the
	only reference, waits while the hardware is deactivated.

Several private helper functions will be defined to simplify the
logic and encapsulate common code.  Here, ipa_client_add_first() is
defined to add the first client under mutex protection.  This
replaces most of the body of ipa3_inc_client_enable_clks().

Change-Id: I71271bbcd84de6fc7659249812d29dcc0d94e086
Signed-off-by: Alex Elder <elder@codeaurora.org>