msm: ipa: restore ipa3_cfg_route()
This commit deleted ipa3_cfg_route(), because it was unused:

  eaf5911 msm: ipa: kill ipa3_cfg_route()

It turns out that it *should* have been used, because it's necessary
to configure the IPA with a default route so exceptional conditions
(bad or unroutable packets) get reported somewhere to be handled
(and we want that "somewhere" to be the AP).  This patch restores
that function in exactly the form it was in when it was removed.
The next patch will use it again.

Change-Id: Iad51f7085f4240f1abf11309ceb5eab970fb4ca7
Signed-off-by: Alex Elder <elder@codeaurora.org>