msm: ipa: now kill setup_default_route()
The only place ipa3_cfg_route() is uses is in setup_default_route().
Change the former function to just set up the default route rather
than any route, and rename it to reflect its updated purpose.  The
only information it needs is the id of the client to which IPA
exceptions should be routed.  This is basically changing merging
the functionality of setup_default_route() into (newly renamed)
ipa_cfg_default_route().

We know the one caller provides a valid client id; assert that in
the function so this requirement is obvious.  As a result there is
no way this function can fail, so give it void return type.

Change-Id: I86ee8a4b18e74cbc42712e03004fcc33420554a9
Signed-off-by: Alex Elder <elder@codeaurora.org>