msm: ipa: don't compute basename so much
Every call to the IPA_ACTIVE_CLIENTS_PREP_*() macros involves the
use of __FILENAME__, which expands to computing the basename of the
__FILE__ path.  (Unfortunately the C preprocessor can't provide this
for us.)

At least for IPA_ACTIVE_CLIENTS_PREP_SIMPLE() this is completely
unnecessary, because such calls never actually add anything to the
log.

Instead, pass __FILE__ to ipa3_active_clients_log_insert() and
compute the basename there--where the log entry is formatted, rather
than at every location that might modify the number of active clients.

Change-Id: I27fff746d66f82e93daa6de1259112b0dc30e41b
Signed-off-by: Alex Elder <elder@codeaurora.org>