msm: ipa: don't print type in active clients table
An "active client" represents a client reference taken from a unique
spot in the code.  Each active client reference has an "id string"
that's used to determine which reference increments and decrements
are associated with it.

Each active client reference has a type associated with it.  The
type is assigned when the first reference is taken, but it is
essentially ignored thereafter.  The type seems to be an easy way
to categorize references (there our four: EP, SIMPLE, RESOURCE, and
SPECIAL), but the form the "id string" takes for each type is distinct
enough that I don't think knowing the type really adds much value.

The id strings for each type are:
  - EP - an endpoint name, like "APPS_CMD_PROD" or "APPS_WAN_CONS"
  - SIMPLE - a function name, like "ipa3_wwan_probe"
  - RESOURCE - a resource name like "RESOURCE_WLAN_PROD"
  - SPECIAL - a special string, like "NAPI" or "PROXY_CLK_VOTE"

Furthermore, the type strings are only used when showing the
(relatively small) active client table; they are not used for the
log, which (currently) contains up to 120 entries.  The small number
of entries in the active client table means that understanding the
type of reference based on context for an entry is not especially
difficult.

This patch just stops printing the type when dumping the contents of
the active client table.  This is in preparation for removing the
notion of a "active client log type" altogether.

Change-Id: I99a6760d8663bba74aeeea921125423776d9df85
Signed-off-by: Alex Elder <elder@codeaurora.org>