msm: ipa: kill active_client_type_string()
The only place active_client_type_string() is used is in an error
message in _ipa3_active_clients_log_mod().  As argued in the
previous patch, the type doesn't add much information when the id
string can be supplied, so there's not much benefit to using it in
this error message.  Remove the type string from the error message,
and remove the function, which becomes unused.

Change-Id: I984be244c46b6e01670ea530d0cad0a7b39b497f
Signed-off-by: Alex Elder <elder@codeaurora.org>