msm: ipa: add Boolean argument to ipa3_active_clients_log_mod()
The only use of the "type" field in the id structure passed to
ipa3_active_clients_log_mod() is to determine whether to add an
entry to the circular log of client reference count changes.
(It is logged for any type other than SIMPLE.) Use a Boolean
argument rather than the type in the id structure to represent
that information.

Change-Id: I00053d97ddbebfa96596a6ac4cade95c253fdd6f
Signed-off-by: Alex Elder <elder@codeaurora.org>