msm: ipa: don't copy id string in active_client_get()
The id string passed via the "id" structure is always a constant and
static string:  for a simple log entry it's __func__; for an endpoint
log entry it's a client name (from the ipa_clients_strings[] array);
and for a special log entry it's a fixed identifier like "NAPI".

These strings are persistent so there's no need to copy them in
active_client_get().  We can just stash a pointer to each string.

Change-Id: I671ebb0178b0b41ef2e5cf05081e730722d8bc5f
Signed-off-by: Alex Elder <elder@codeaurora.org>