msm: ipa: simplify ipa3_get_bus_vote()
The curr_ipa_clk_rate field of the ipa3_context structure is
assigned value 0 in ipa3_pre_init().  It is never assigned
thereafter.  We know therefore that it will be 0 when
ipa3_get_bus_vote() is called, and in that case the value
that function should return is 1.  Delete most of the code in
ipa3_get_bus_vote() to make it obvious what happens.

Change-Id: I420d88d2db52b484b60142af01da15ba51c27e51
Signed-off-by: Alex Elder <elder@codeaurora.org>