msm: ipa: rename former controller functions
All but one of the functions that were previously accessed through
the IPA controller structure are defined in the same file they're
used, so they can be given static scope.  Rename these, dropping the
underscore prefix, and removing the version number.  When we have
something that supports other version we can re-implement versioning
in the name.

Change-Id: I61998a1295be1303cab7f10421b6f648d82b1228
Signed-off-by: Alex Elder <elder@codeaurora.org>