msm: ipa: kill ipa3_mem()
The IPA memory array is now in the IPA context, not the controller
structure.  The accessor function ipa3_mem() is trivial, so get rid
of that and just access entries in the ipa3_ctx->mem_info[] array
directly.  This change amounts to a global substitution of the one
access method for the other.

Change-Id: Ifb2ffe390db0a0c6f8a37eb0691a9bc088dcd4ed
Signed-off-by: Alex Elder <elder@codeaurora.org>