msm: ipa: introduce sram_set_canaries()
Create a function that sets a pair of consecutive canary values
before an IPA memory range rather than just one, and use it in
ipa_init_sram().  Use a common function to handle both the one and
two canary case, and rename the functions (they're private and don't
really need an IPA prefix).

Change-Id: I5ab6a0ecfc9bffa8844957be5a71eeef6a909a20
Signed-off-by: Alex Elder <elder@codeaurora.org>