msm: ipa: kill struct ipa3_controller
The only remaining field used in the ipa3_controller structure is
the bus scale table.  Move that into the IPA context, and initialize
that field directly with a new function ipa_bus_scale_table_init().
With that we can get rid of the "ctrl" field in the IPA context, and
the ipa3_controller structure and ipa3_controller_init().  Delay
initializing the bus scale table pointer until just before it's
needed for the call to msm_bus_scale_register_client().

Change-Id: I1b30895f87b734ea4836e70b3eccba8fef6f95b8
Signed-off-by: Alex Elder <elder@codeaurora.org>