msm: ipa: fix copy_to_user() size argument
The latest upstream kernel implements more stringent checks of the
size argument provided to copy_to_user().  If it is unable to prove
that the size is within range of the target address it issues a
warning at compile time.

We can fix this by just ensuring the copied size is less than the
size of the target.  We were already checking this, but by limiting
the size copied the compiler can have the information it needs to
avoid the warning.

Implement this fix in two places.

Change-Id: I23c4beb634a7b834b4db1ef8a2b174faf18ed141
Signed-off-by: Alex Elder <elder@codeaurora.org>