msm: ipa: don't register for modem SSR notifications
Don't register with the subsystem-restart facility for notifications
about modem events.  This makes the ipa3_ssr_notifier block unused
so it can be removed; doing that makes ipa3_ssr_notifier_cb() no
longer used so that too can be deleted.  That function was the only
user of ipa_stop_polling_stats(), so that function goes away too.

Change-Id: I0497bddde623e2c171e2e4f6a297570d53101cda
Signed-off-by: Alex Elder <elder@codeaurora.org>