msm: ipa: kill ipa3_tag_process()
The previous patch removed ipa3_q6_set_ex_path_to_apps(), and that
was the only user of ipa3_tag_process(), so that function can now
be removed.  That makes ipa3_destroy_imm() and ipa3_tag_free_skb()
unused, along with a few other constants, so they can be removed as
well.

Change-Id: I5d4767e7ad8f3180e0600f2efc516dda7673596b
Signed-off-by: Alex Elder <elder@codeaurora.org>