msm: ipa: call ipa3_wwan_init() directly
When ipa3_wwan_init() is called via late_initcall(), it is likely
that the IPA will not be in READY state.  Currently the code returns
EPROBE_DEFER in this case, to request for re-probe.  However, the
thing that changes the state of IPA to READY is in user space
(input sent to /dev/ipa), and the device core has no way of knowing
that this is a reason to attempt a re-probe.  This is the reason the
callback registration mechanism that was originally in place was
defined.

This patch implements a simpler solution.  Rather than having
ipa3_wwan_init() get called as a late initcall, we just have
ipa3_write() (which transiations the IPA to READY state) call
ipa3_wwan_init() directly.  This avoids any need for synchronization
between these two platform drivers.  The probe function can then
assume no synchronization is required, so doesn't need to check
whether the IPA driver is ready.

Change-Id: I3c4bfff5e0c710e336ae3279b76f5782b425044c
Signed-off-by: Alex Elder <elder@codeaurora.org>