msm: ipa: introduce UPSTREAM and UPSTREAM_SMEM
Define a symbol UPSTREAM that's used to separate code that's built
for the upstream environment from that which should be built for
"downstream."  The value is based on the kernel version.  This use
of #ifdef's is intended to allow ongoing testing on the same code
base while things get ported to the upstream environment.

Define UPSTREAM_SMEM to surround code that's dependent on the
different ways Qualcomm shared memory is implemented in the two
environments.

Change-Id: I566969d04f0c3b31119d05a1e0f566ef5402732a
Signed-off-by: Alex Elder <elder@codeaurora.org>