msm: ipa: kill ipa3_context->state
The state variable was primarily used to signal to the WWAN driver
that the main IPA code has finished initializing.  The previous
patch made that unnecessary.  It's still set, and when initally set
it was set to STARTING (but that was primarily as a convenience).
So get rid of the "state" field, as well as the definition of the
values it takes on.

Change-Id: Ibddf2ecb3665601b421299ae9d79d4cc7a55c5ac
Signed-off-by: Alex Elder <elder@codeaurora.org>