msm: ipa: set up timer for no-op requests
Creeate three functions used in managing a high-resolution timer
that arranges for a no-op request to be sent after transmit request
has been issued to a producer pipe configured for deferred
interrupt handling.
  - ipa_nop_timer_init() initializes the timer state for a pipe
    marked to not have transmit requests interrupt.
  - ipa_nop_timer_schedule() is used to start the timer if it is not
    already running.  An atomic variable is used to ensure the timer
    is always started if one is not pending.
  - ipa_nop_timer_expiry() is the timer callback function, called
    when the timer expires.  It arranges for a no-op request to be
    sent to the pipe in workqueue context.

The timeout period was never deleted before; keep it at 2 msec, but
give it a shorter name.

Note:  despite the atomic variable, there is still a race here that
could lead to transmit requests never completing.  If the hrtimer
code supported a "start if not already running" function that did
not update (extend) the timeout we could use that.  It seems we
aren't suffering from this problem, so we'll delay addressing it.

Change-Id: I233c468388d1eccba4ef0949115cbef55596ec8b
Signed-off-by: Alex Elder <elder@codeaurora.org>