msm: ipa: introduce ipa_send_nop()
There previously was a "common event ring" that was shared by
producer pipes.  That event ring had two features:  it allowed
multiple producer pipes to share the same event ring for signaling
transmit completions; and it arranged for requests using that event
ring to not interrupt.  Instead of interrupting on every request
completion, an interrupting no-op request is sent after a short
delay to handle completions in batches.

This patch begins to add back the second feature.  Only the WAN
producer pipe will use this capability, so there's no real benefit
to setting up a special "shared" event ring.  But we can still
benefit from deferring interrupt handling.

This patch introduces ipa_send_nop(), which issues a no-op request
that will interrupt when it completes.  This interrupt will be what
causes accumulated transmit completions to be handled.  The no-op
request has no data, so the virtual address associated with it will
be NULL.  The previous patch arranged for ipa3_wq_write_done() to
handle that properly.

For now, this function isn't used; add a reference to it to avoid a
compiler error.

Change-Id: I17dae1be27c306de6be896c55278e45764568198
Signed-off-by: Alex Elder <elder@codeaurora.org>