ipa: qmi: fix some things before removing old QMI code
A few things need to be rearranged a bit before we can remove the
old IPA QMI code.  The main thing is that the definition of the
ipa3_rmnet_mux_val structure must be moved into "rmnet_ipa.c" so
it's available there (not used elsewhere).  That file also needs
to include "ipa_i.h"; it required it before, but the dependency
was satisfied through the inclusion of "ipa_qmi_service.h".

We are now assuming we have an Android platform, so drop the
platform type argument from ipa3_qmi_service_init().  Rename it,
as well as its _exit counterpart.

Signed-off-by: Alex Elder <elder@linaro.org>