ipa: port to upstream SMMU/IOMMU interfaces
This implements changes comparable to the patch supplied to me by
Abhishek:
    msm: ipa: IPAlite fastmap changes
It modifies the SMMU code to be compatible with the upstream
environment.

We no longer keep an IOMMU mapping in the SMMU context block,
instead saving an IOMMU domain.  Also, ipa_smmu_domain_attr_set()
is now gone, because we only set the FAST attribute on the domain
(and that is now done in a conditional block, awaiting the fastmap
capability being available).

Signed-off-by: Alex Elder <elder@linaro.org>