soc: qcom: ipa: zip table info is invalid in QMI message
When we formulate the QMI INIT_MODEM_DRIVER message to send to the
modem in init_modem_driver_req(), we know the size of the modem
compression/decompression region is zero.  As a result, the table
info is not valid, and won't be sent with the message.

The fields in that QMI structure related to the zip table can all
just be zero.  Because the structure is static, it will be initially
zero, so we can get away with just not assigning those fields.

Signed-off-by: Alex Elder <elder@linaro.org>