soc: qcom: ipa: remove NAT-related memory info
The size of the NAT memory region is zero, and is furthermore
unused.  Someday we can re-enable this IPA feature, but in the mean
time we don't need to keep track of it.  So get rid of the NAT_OFST
and NAT_SIZE index values and remove the slots in the mem_info array
for them.

Signed-off-by: Alex Elder <elder@linaro.org>