soc: qcom: ipa: kill ipa3_init_smem_region()
Now that ipa3_init_smem_region() is simply a wrapper around a call
to dma_shared_mem_zero_cmd(), replace the use of the former with the
latter.  The only thing missing from dma_shared_mem_zero_cmd() is an
assertino that the size provided is zero.  All callers currently
provide non-zero size values, so we can safely add that sensible
assertion there as well.

Signed-off-by: Alex Elder <elder@linaro.org>