soc: qcom: ipa: kill rmnet_ipa3_ctx->is_initialized
The is_initialized field of the rmnet_ipa3_ctx structure is set
atomically to 0 or 1 to indicate when initialization has completed
successfully.

It is set in the probe and release functions (and ipa3_wwan_init(),
which is expected to be called before the probe).  The probe and
release functions are protected by a mutex, so there is no need for
this to be an atomic.  Furthermore, nothing ever checks its value,
so it's really not required.

On the other hand, it is called by ipa3_post_init(), which currently
gets called as a result of a write to /dev/ipa, which could
conceivably be done more than once.

This patch removes the is_initialized field in the IPA RMnet context.
However it creates a new Boolean global variable "initialized" that
is set and cleared at the same time the is_initialized field was
previously.  There are two differences now:  the variable is no
longer atomic; and it is tested at the top of ipa3_wwan_init() to
avoid performing initialization more than once.

Signed-off-by: Alex Elder <elder@linaro.org>